+++ shopit/cms_apps.py
@@ -6,7 +6,13 @@ from cms.apphook_pool import apphook_pool
 from django.utils.translation import ugettext_lazy as _
 
 from shopit.settings import SINGLE_APPHOOK
-from shopit.urls import get_urls
+
+
+# TODO: can be simplified into, after merging https://github.com/divio/django-cms/pull/5898
+# from shopit.urls import get_urls
+
+# def get_urls(self, page=None, language=None, **kwargs):
+#     return get_urls('module')
 
 
 class ShopitApphook(CMSApp):
@@ -15,42 +21,42 @@ class ShopitApphook(CMSApp):
     def get_urls(self, page=None, language=None, **kwargs):
         if SINGLE_APPHOOK:
             return ['shopit.urls']
-        return [get_urls('shop')]
+        return ['shopit.urls.shop']
 
 
 class ShopitAccountApphook(CMSApp):
     name = _('Shopit Account')
 
     def get_urls(self, page=None, language=None, **kwargs):
-        return [get_urls('account')]
+        return ['shopit.urls.account']
 
 
 class ShopitCategoriesApphook(CMSApp):
     name = _('Shopit Categories')
 
     def get_urls(self, page=None, language=None, **kwargs):
-        return [get_urls('categories')]
+        return ['shopit.urls.categories']
 
 
 class ShopitBrandsApphook(CMSApp):
     name = _('Shopit Brands')
 
     def get_urls(self, page=None, language=None, **kwargs):
-        return [get_urls('brands')]
+        return ['shopit.urls.brands']
 
 
 class ShopitManufacturersApphook(CMSApp):
     name = _('Shopit Manufacturers')
 
     def get_urls(self, page=None, language=None, **kwargs):
-        return [get_urls('manufacturers')]
+        return ['shopit.urls.manufacturers']
 
 
 class ShopitProductsApphook(CMSApp):
     name = _('Shopit Products')
 
     def get_urls(self, page=None, language=None, **kwargs):
-        return [get_urls('products')]
+        return ['shopit.urls.products']
 
 
 apphook_pool.register(ShopitApphook)