+++ shopit/urls/products.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from shopit.urls import get_urls
+
+
+urlpatterns = get_urls('products')